use graphql fetcher by default
